fix: widen loopCount to long and use floorMod for wal index

+++ src/main/java/io/velo/task/TaskChain.java
@@ -37,7 +37,7 @@ public class TaskChain {
      *
      * @param loopCount the current loop count
      */
-    public void doTask(int loopCount) {
+    public void doTask(long loopCount) {
         for (var t : list) {
             if (loopCount % t.executeOnceAfterLoopCount() == 0) {
                 t.setLoopCount(loopCount);
